requirements/dev.txt
# you will need to install these requirements with `pip install -r requirements.txt`
aiohttp-devtools==0.11
uvloop==0.12.0